Project 2: Initial Commit
I'm gonna use the same repository to start the next project of this course, which consists in creating a store based on the product used on the first project. I'm gonna concentrate in HTML first and then i'll start with SCSS and finally, JS.

I created a new folder where all the HTML files are going to be. Also, I linked the Valorant Landing Page with the Skin Store Page.