More HTML files
This are more or less empty too, but, there has been a notable progress today. I'm doing this step by step, slowly. The plan is to focus 100% on friday. I also have to make changes in the store design, because there are some thing that make no sense, so, i'll be doing that until friday.