Firestore, Login, Logout & Register are Functional.
The project is already linked with Firestore. Registration and login are functional.

When a user logs in, the buttons on the navigation menu are replaced with the name the user entered when registering and a button to log out. After the user logs out, the Login and Register buttons reappear.